RE-179 fix: folder res type Long

+++ src/main/java/info/reinput/reinputworkspaceservice/folder/presentation/dto/res/FolderRes.java
@@ -9,7 +9,7 @@ public record FolderRes (
         Long folderId,
         String folderName,
         String folderColor,
-        Integer insightCount,
+        Long insightCount,
         @JsonInclude(JsonInclude.Include.NON_NULL)
         ShareRes share
 ){
